Spotify_Intelligence/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
spotipy>=2.23.0
xgboost>=2.0.0
shap>=0.45.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.20.0
google-genai>=0.8.0
python-dotenv>=1.0.0
scipy>=1.12.0
requests>=2.32.0
